safari-600.7-branch/LayoutTests/platform/ios-sim/http/tests/security/cross-origin-appcache-expected.txt

This iframe should not have a cache:

This iframe should have a cache:



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Cache not found

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Cache found
